fw/drivers/pmic/npm1300: fixes for C0 hw charging issues
C0 hw (used in asterix), may have some issues detecting batteries that
report low voltage (due to e.g. PCM disconnecting the cell). Add some
workarounds so that C0 behaves similar to D0, ie, battery detection is
removed and not a condition to start charging.

We may apply this conditionally if other boards use D0, but it's really
a no-op on D0 so it does not harm.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>