third_party/memfault: fix MEMFAULT define not being set globally
The MEMFAULT preprocessor define was moved to the build function in
64837143, which meant it was only set in the local memfault build
context and not propagated to firmware sources. This caused all
`#if MEMFAULT` checks in src/fw/ to evaluate as undefined/false,
breaking memfault metrics collection for release builds.

Move the define back to configure() so it's set globally for all build
targets, while keeping the Obelix PRF library build disabled.

Signed-off-by: Joshua Jun <joshuajun@proton.me>